requirements.txt
google-generativeai==0.8.5
gradio==5.35.0
langchain-google-vertextai==2.0.27
langchain==0.3.26
pydantic==2.11.7
python==3.10.18
python-dotenv==1.1.1
Pillow==11.3.0
opencv-python==4.12.0.88
PyMuPDF==1.26.3
watchdog==6.0.0